\docType{methods}
\name{hits}
\alias{getId}
\alias{hits}
\title{Accessor methods for blast records}
\arguments{
  \item{x}{A \code{\link{blastReccord-class}} object.}

  \item{...}{Additional arguments}

  \item{x}{A blastReport object.}

  \item{db}{Database tag (e.g.: 'gi', 'gb', 'emb', 'ref',
  ...)}
}
\value{
  Accession numbers as specified by \code{db} for each hit.
}
\description{
  Accessor methods for blast records

  Accessor methods for blast records
}

