% Built by roxygen3 0.1
\docType{methods}
\name{c,efetch-method}
\alias{c,efetch-method}
\title{Combining efetch objects}
\usage{
  c(...)
}
\arguments{
  \item{...}{objects to be combined.}
}

\description{
Only data retrieved from the same datebase in the same \code{retmode} and
\code{rettype} can be combined.
}
\keyword{methods}
