\docType{methods}
\name{docsum}
\alias{docsum}
\title{Access a DocSum from an \code{\link{esummary-class}} object.}
\arguments{
  \item{object}{An \code{\link{esummary-class}} object.}
}
\value{
  A data frame or a list
}
\description{
  Attempts to parse ESummary DocSums into a data frame.
  Returns a named list if not implemented for a specific
  database.
}
\examples{
## examples
}

