\name{egquery}
\alias{egquery}
\title{Retrieve the number of records in all Entrez databases for a single
text query}
\usage{
  egquery(term)
}
\arguments{
  \item{term}{A valid Entrez text query.}
}
\value{
  An \code{\link{egquery-class}} object.
}
\description{
  \code{egquery} provides Entrez database counts for a
  global search.
}
\details{
  See the online documentation at
  \url{http://www.ncbi.nlm.nih.gov/books/NBK25499/#chapter4.EGQuery}
  for additional information.
}

