% Built by roxygen3 0.1
\name{dbName}
\alias{dbName}
\alias{description}
\alias{fields}
\alias{lastUpdate}
\alias{links}
\alias{menuName}
\alias{records}
\title{EInfo accessors}
\usage{
  dbName(x, ...)

  menuName(x, ...)

  description(x, ...)

  records(x, ...)

  lastUpdate(x, ...)

  fields(x, ...)

  links(x, ...)
}
\arguments{
  \item{x}{An \code{\linkS4class{einfo}} instance.}

  \item{...}{Further arguments passed on to methods.}
}

\description{
Access the various slots of \linkS4class{einfo} instances.
}
\section{Method signatures for generic dbName}{
\itemize{
  \item \linkS4class{einfo}
}

}

\section{Method signatures for generic menuName}{
\itemize{
  \item \linkS4class{einfoDb}
}

}

\section{Method signatures for generic description}{
\itemize{
  \item \linkS4class{einfoDb}
}

}

\section{Method signatures for generic records}{
\itemize{
  \item \linkS4class{einfoDb}
}

}

\section{Method signatures for generic lastUpdate}{
\itemize{
  \item \linkS4class{einfoDb}
}

}

\section{Method signatures for generic fields}{
\itemize{
  \item \linkS4class{einfoDb}
}

}

\section{Method signatures for generic links}{
\itemize{
  \item \linkS4class{einfoDb}
}

}

\keyword{methods}
