\name{einfo-class}
\alias{einfo-class}
\alias{einfo,einfo-method}
\alias{show,einfo-method}
\title{einfo class}
\description{
  einfo is a virtual S4 class that is extended by the
  \code{\link{einfoDbList-class}}, and
  \code{\link{einfoDb-class}} classes.
}
\seealso{
  \code{\link{einfo}} for generating calls to the NCBI
  EInfo utility.
}

