% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_info.r
\name{entrez_dbs}
\alias{entrez_dbs}
\title{List databases available from the NCBI}
\usage{
entrez_dbs(config = NULL)
}
\arguments{
\item{config}{config vector passed to \code{httr::GET}}
}
\value{
character vector listing available dbs
}
\description{
Retrieves the names of  databases available through the EUtils API
}
\examples{
\dontrun{
entrez_dbs()
}
}
\seealso{
Other einfo:
\code{\link[=entrez_db_links]{entrez_db_links()}},
\code{\link[=entrez_db_searchable]{entrez_db_searchable()}},
\code{\link[=entrez_db_summary]{entrez_db_summary()}},
\code{\link[=entrez_info]{entrez_info()}}
}
\concept{einfo}
