% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/set_globals.r
\docType{data}
\name{entrez_email}
\alias{entrez_email}
\title{Set global variables}
\format{\preformatted{ chr "david.winter@gmail.com"
}}
\usage{
entrez_email
}
\description{
Set global variables
}
\keyword{datasets}

