% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_global_query.r
\name{entrez_global_query}
\alias{entrez_global_query}
\title{Find the number of records that match a given term across all NCBI Entrez databases}
\usage{
entrez_global_query(term, config = NULL, retry = entrez_retry_options(), ...)
}
\arguments{
\item{term}{the search term to use}

\item{config}{vector configuration options passed to httr::GET}

\item{retry}{(list) use \code{\link{entrez_retry_options}} to make a named
list of retry options to pass on to the HTTP request. default values are
passed for you, but you can change them by setting the params in
\code{\link{entrez_retry_options}} passed to this parameter}

\item{...}{additional arguments to add to the query}
}
\value{
a named vector with counts for each a database
}
\description{
Find the number of records that match a given term across all NCBI Entrez databases
}
\examples{
\dontrun{ 
NCBI_data_on_best_butterflies_ever <- entrez_global_query(term="Heliconius")
}
}
\seealso{
\code{\link[httr]{config}} for available configs
}
