% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.r
\name{entrez_retry_options}
\alias{entrez_retry_options}
\title{entrez retry options}
\usage{
entrez_retry_options(
  pause_base = 1,
  pause_cap = 60,
  pause_min = 1,
  times = 3,
  terminate_on = NULL,
  terminate_on_success = TRUE
)
}
\arguments{
\item{pause_base, pause_cap, pause_min}{basis, maximum, and minimum for
calculating wait time for retry.}

\item{times}{the maximum number of times to retry.}

\item{terminate_on}{a vector of HTTP status codes.}

\item{terminate_on_success}{If `TRUE`, the default, this will
automatically terminate when the request is successful, regardless of the
value of `terminate_on`}
}
\value{
a named list with the parameters given to this function
}
\description{
entrez retry options
}
\details{
see [httr::RETRY] for more detailed explanation of these
parameters
}
