% Built by roxygen3 0.1
\name{error}
\alias{error}
\title{Extract errors from an eutil request.}
\usage{
  error(x, ...)
}
\arguments{
  \item{x}{An \code{\linkS4class{eutil}} object}

  \item{...}{Further arguments passed on to methods.}
}

\description{
Extract errors from an eutil request.
}
\section{Method signatures for generic error}{
\itemize{
  \item \linkS4class{eutil}
}

}

\keyword{methods}
