\name{.query}
\alias{.httpPOST}
\alias{.query}
\title{Construct url, fetch response, construct eutil object}
\usage{
  .query(eutil, ...)

  .httpPOST(eutil, ...)
}
\arguments{
  \item{eutil}{Which eutil?}

  \item{...}{Additional args.}
}
\description{
  Construct url, fetch response, construct eutil object

  use HTTP POST
}
\keyword{internal}

