% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entrez_link.r
\name{linkout_urls}
\alias{linkout_urls}
\title{Extract urls from an elink object}
\usage{
linkout_urls(elink)
}
\arguments{
\item{elink:}{elink object (returned by entrez_link) containing Urls}

\item{list}{of character vectors, one per ID each containing of urls for that
ID.}
}
\description{
Extract urls from an elink object
}
\seealso{
entrez_link
}

