% Built by roxygen3 0.1
\name{query}
\alias{query}
\title{Extract url from an eutil request.}
\usage{
  query(x, ...)
}
\arguments{
  \item{x}{An \code{\linkS4class{eutil}} object}

  \item{...}{Further arguments passed on to methods.}
}

\description{
Extract url from an eutil request.
}
\section{Method signatures for generic query}{
\itemize{
  \item \linkS4class{eutil}
}

}

\keyword{methods}
