% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rentrez-package.R
\docType{package}
\name{rentrez-package}
\alias{rentrez}
\alias{rentrez-package}
\title{rentrez: 'Entrez' in R}
\description{
Provides an R interface to the NCBI's 'EUtils' API, allowing users to search databases like 'GenBank' \url{https://www.ncbi.nlm.nih.gov/genbank/} and 'PubMed' \url{https://pubmed.ncbi.nlm.nih.gov/}, process the results of those searches and pull data into their R sessions.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/rentrez/}
  \item \url{https://github.com/ropensci/rentrez/}
  \item Report bugs at \url{https://github.com/ropensci/rentrez/issues/}
}

}
\author{
\strong{Maintainer}: Samuel Bharti \email{samuelbharti.io@gmail.com} (\href{https://orcid.org/0000-0003-4190-7058}{ORCID})

Authors:
\itemize{
  \item Samuel Bharti \email{samuelbharti.io@gmail.com} (\href{https://orcid.org/0000-0003-4190-7058}{ORCID})
  \item David Winter \email{david.winter@gmail.com} (\href{https://orcid.org/0000-0002-6165-0029}{ORCID})
  \item Indraneel Chakraborty \email{hello.indraneel@gmail.com} (\href{https://orcid.org/0000-0001-6958-8269}{ORCID})
}

Other contributors:
\itemize{
  \item Scott Chamberlain \email{myrmecocystus@gmail.com} (\href{https://orcid.org/0000-0003-1444-9135}{ORCID}) [contributor]
  \item Han Guangchun \email{hanguangchun@gmail.com} (\href{https://orcid.org/0000-0001-9277-2507}{ORCID}) [contributor]
}

}
\keyword{internal}
