% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_keys.r
\name{set_entrez_key}
\alias{set_entrez_key}
\title{Set the ENTREZ_KEY variable to be used by all rentrez functions}
\usage{
set_entrez_key(key)
}
\arguments{
\item{key}{character. Value to set ENTREZ_KEY to (i.e. your API key).}
}
\value{
A logical of length one, TRUE is the value was set FALSE if not.
value is returned inside invisible(), i.e. it is not printed to screen 
when the function is called.
}
\description{
The NCBI allows users to access more records (10 per second) if they 
register for and use an API key. This function allows users to set this key
for all calls to rentrez functions during a particular R session. See the
vignette section "Using API keys" for a detailed description.
}
