\name{show,einfo-method}
\alias{show,efetch-method}
\alias{show,egquery-method}
\alias{show,einfo-method}
\alias{show,elink-method}
\alias{show,epost-method}
\alias{show,esearch-method}
\alias{show,esummary-method}
\title{method extensions to show for eutil objects.}
\usage{
  \S4method{show}{esearch}(object)

  \S4method{show}{esummary}(object)

  \S4method{show}{efetch}(object)

  \S4method{show}{epost}(object)

  \S4method{show}{elink}(object)

  \S4method{show}{egquery}(object)
}
\arguments{
  \item{object}{Any R object}
}
\description{
  See the documentation of \code{\link[methods]{show}}
  method for the expected behavior.
}
\seealso{
  \code{\link[methods]{show}}
}

