\docType{methods}
\name{content}
\alias{content}
\title{Extract content from an eutil request.}
\arguments{
  \item{x}{an \code{\link{eutil-class}} object.}

  \item{parse}{Parse into an R object where possible.}

  \item{format}{Output format of sequence data.}

  \item{...}{Further arguments.}
}
\description{
  Retrieves the contents of an eutil request either as is,
  or attempts to return an R object.
}

