\name{einfo-class}
\alias{einfo-class}
\title{\dQuote{einfo} class}
\description{
  einfo is a virtual S4 class that is extended by the
  \code{\linkS4class{einfoDbList}}, and
  \code{\linkS4class{einfoDb}} classes.
}
\seealso{
  \code{\link{einfo}} for generating calls to the NCBI
  EInfo utility.
}

