% Generated by roxygen2 (4.0.0): do not edit by hand
\name{entrez_db_links}
\alias{entrez_db_links}
\title{Can be used in conjunction with \code{\link{entrez_link}} to find
the right name for the \code{db_to} argument}
\usage{
entrez_db_links(db, config = NULL)
}
\arguments{
\item{config}{config vector passed to \code{httr::GET}}

\item{db}{character, name of database t}
}
\value{
An eInfoLink object (subclassed from list) summarising linked-datbases.
Can be coerced to a data-frame with \code{as.data.frame}. Printing the object
the name of each element.
}
\description{
Can be used in conjunction with \code{\link{entrez_link}} to find
the right name for the \code{db_to} argument
}
\seealso{
\code{\link{entrez_link}}

Other einfo: \code{\link{entrez_db_searchable}};
  \code{\link{entrez_db_summary}};
  \code{\link{entrez_dbs}}; \code{\link{entrez_info}}
}

