% Generated by roxygen2 (4.0.0): do not edit by hand
\name{entrez_db_searchable}
\alias{entrez_db_searchable}
\title{Can be used in conjunction with \code{\link{entrez_search}} to find avaliable
search fields to include in the \code{term} argument}
\usage{
entrez_db_searchable(db, config = NULL)
}
\arguments{
\item{config}{config vector passedto \code{httr::GET}}

\item{db}{character, name of database to get search field from}
}
\value{
An eInfoSearch object (subclassed from list) summarising linked-datbases.
Can be coerced to a data-frame with \code{as.data.frame}. Printing the object
shows only the names of each avaliable search field.
}
\description{
Can be used in conjunction with \code{\link{entrez_search}} to find avaliable
search fields to include in the \code{term} argument
}
\seealso{
\code{\link{entrez_search}}

Other einfo: \code{\link{entrez_db_links}};
  \code{\link{entrez_db_summary}};
  \code{\link{entrez_dbs}}; \code{\link{entrez_info}}
}

