% Generated by roxygen2 (4.0.0): do not edit by hand
\name{entrez_db_summary}
\alias{entrez_db_summary}
\title{Summarise a NCBI database}
\usage{
entrez_db_summary(db, config = NULL)
}
\arguments{
\item{config}{config vector passedto \code{httr::GET}}

\item{db}{character, name of database t}
}
\value{
Character vector with the following data

DbName Name of database

Description Brief description of the database

Count Number of records contained in the databse

MenuName Name in web-interface to EUtils

DbBuild Unique ID for current build of databse

LastUpdate Date of most recent update to databse
}
\description{
Summarise a NCBI database
}
\seealso{
Other einfo: \code{\link{entrez_db_links}};
  \code{\link{entrez_db_searchable}};
  \code{\link{entrez_dbs}}; \code{\link{entrez_info}}
}

