% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entrez_info.r
\name{entrez_db_summary}
\alias{entrez_db_summary}
\title{Retrieve summary information about an NCBI database}
\usage{
entrez_db_summary(db, config = NULL)
}
\arguments{
\item{db}{character, name of database to summaries}

\item{config}{config vector passed to \code{httr::GET}}
}
\value{
Character vector with the following data

DbName Name of database

Description Brief description of the database

Count Number of records contained in the database

MenuName Name in web-interface to EUtils

DbBuild Unique ID for current build of database

LastUpdate Date of most recent update to database
}
\description{
Retrieve summary information about an NCBI database
}
\examples{
entrez_db_summary("pubmed")
}
\seealso{
Other einfo: \code{\link{entrez_db_links}};
  \code{\link{entrez_db_searchable}};
  \code{\link{entrez_dbs}}; \code{\link{entrez_info}}
}

