% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_info.r
\name{entrez_db_summary}
\alias{entrez_db_summary}
\title{Retrieve summary information about an NCBI database}
\usage{
entrez_db_summary(db, config = NULL, retry = entrez_retry_options())
}
\arguments{
\item{db}{character, name of database to summaries}

\item{config}{config vector passed to \code{httr::GET}}

\item{retry}{(list) use \code{\link{entrez_retry_options}} to make a named
list of retry options to pass on to the HTTP request. default values are
passed for you, but you can change them by setting the params in
\code{\link{entrez_retry_options}} passed to this parameter}
}
\value{
Character vector with the following data

DbName Name of database

Description Brief description of the database

Count Number of records contained in the database

MenuName Name in web-interface to EUtils

DbBuild Unique ID for current build of database

LastUpdate Date of most recent update to database
}
\description{
Retrieve summary information about an NCBI database
}
\examples{
\dontrun{
entrez_db_summary("pubmed")
}
}
\seealso{
Other einfo: 
\code{\link{entrez_db_links}()},
\code{\link{entrez_db_searchable}()},
\code{\link{entrez_dbs}()},
\code{\link{entrez_info}()}
}
\concept{einfo}
