% Generated by roxygen2 (4.0.0): do not edit by hand
\name{entrez_dbs}
\alias{entrez_dbs}
\title{List databases avaliable from the NCBI}
\usage{
entrez_dbs(config = NULL)
}
\arguments{
\item{config}{config vector passedto \code{httr::GET}}
}
\value{
character vector listing avaliable dbs
}
\description{
Retreives a list of databases avaliable from NCBI
}
\examples{
\donttest{
entrez_dbs()
}
}
\seealso{
Other einfo: \code{\link{entrez_db_links}};
  \code{\link{entrez_db_searchable}};
  \code{\link{entrez_db_summary}};
  \code{\link{entrez_info}}
}

