% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/entrez_info.r
\name{entrez_dbs}
\alias{entrez_dbs}
\title{List databases available from the NCBI}
\usage{
entrez_dbs(config = NULL, retry = entrez_retry_options())
}
\arguments{
\item{config}{config vector passed to \code{httr::GET}}

\item{retry}{(list) use \code{\link{entrez_retry_options}} to make a named
list of retry options to pass on to the HTTP request. default values are
passed for you, but you can change them by setting the params in
\code{\link{entrez_retry_options}} passed to this parameter}
}
\value{
character vector listing available dbs
}
\description{
Retrieves the names of  databases available through the EUtils API
}
\examples{
\dontrun{
entrez_dbs()
}
}
\seealso{
Other einfo: 
\code{\link{entrez_db_links}()},
\code{\link{entrez_db_searchable}()},
\code{\link{entrez_db_summary}()},
\code{\link{entrez_info}()}
}
\concept{einfo}
