% Generated by roxygen2 (4.0.0): do not edit by hand
\name{entrez_info}
\alias{entrez_info}
\title{Get infromation about EUtils databases}
\usage{
entrez_info(db = NULL, config = NULL)
}
\arguments{
\item{db}{characater database about which to retreive information (optional,}

\item{config}{config vector passed on to \code{httr::GET}}
}
\value{
XMLInternalDocument with information describing either all the
databases avaliable in Eutils (if db is not set) or one pariticular database
(set by 'db')
}
\description{
Contstructs a query to NCBI's einfo and return parsed XML object
Note The most common uses for einfo API are to list fields against which
a database can be searched, or linked databases avaliabl,
}
\examples{
all_the_data <- entrez_info()
xpathSApply(all_the_data, "//DbName", xmlValue)
}
\seealso{
\code{\link[httr]{config}} for avaliable httr configurations

Other einfo: \code{\link{entrez_db_links}};
  \code{\link{entrez_db_searchable}};
  \code{\link{entrez_db_summary}}; \code{\link{entrez_dbs}}
}

