% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entrez_summary.r
\name{extract_from_esummary}
\alias{extract_from_esummary}
\title{Extract elements from a list of esummary records}
\usage{
extract_from_esummary(esummaries, elements, simplify = TRUE)
}
\arguments{
\item{esummaries}{A list of esummary objects}

\item{elements}{the names of the element to extract}

\item{simplify}{logical, if possible return a vector}
}
\value{
List or vector containing requested elements
}
\description{
Extract elements from a list of esummary records
}

