% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/entrez_summary.r
\name{extract_from_esummary}
\alias{extract_from_esummary}
\title{Extract elements from a list of esumarrry records}
\usage{
extract_from_esummary(esummaries, elements, simplify = TRUE)
}
\arguments{
\item{esummaries}{A list of esummary objects}

\item{elements}{the names of the element to extract}

\item{logical,}{if possible return a vector}
}
\description{
Extract elements from a list of esumarrry records
}

