% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/entrez_link.r
\name{linkout_urls}
\alias{linkout_urls}
\title{Extract URLs from an elink object}
\usage{
linkout_urls(elink)
}
\arguments{
\item{elink}{elink object (returned by entrez_link) containing Urls}
}
\value{
list of character vectors, one per ID each containing of URLs for that
ID.
}
\description{
Extract URLs from an elink object
}
\seealso{
entrez_link
}

