\name{parse_pubmed_xml}
\alias{parse_pubmed_xml}
\title{Summarise an XML record from pubmed.}
\usage{
  parse_pubmed_xml(paper)
}
\arguments{
  \item{paper}{character the record to be parsed (as a
  character, expected to come from
  \code{\link{entrez_fetch}})}
}
\description{
  Summarise an XML record from pubmed.
}

